#259 - Refactor variable mapping functions and update imports in Environment and Settings components
